fix: add section button reloading page
