+++ src/components/proposal/Form.tsx
@@ -3,22 +3,15 @@
 import Header from "./Header";
 import InfoSection from "./InfoSection";
 import GPTGeneratedSection from "./GPTGeneratedSection";
-import { useState } from "react";
 import DeleteButton from "./buttons/DeleteButton";
 import Row from "./Row";
-import AddButton from "./buttons/AddButton";
 
-export default function Form() {
-  const [addSection, setAddSection] = useState(1);
+interface FormProps {
+  onClickDelete: () => void;
+  sectionsCount: number;
+}
 
-  const handleAddSection = () => {
-    setAddSection((prev) => prev + 1);
-  };
-
-  const handleRemoveSection = () => {
-    setAddSection((prev) => prev - 1);
-  };
-
+export default function Form({ onClickDelete, sectionsCount }: FormProps) {
   return (
     <form className="flex flex-col gap-8">
       <Row>
@@ -27,18 +20,12 @@ export default function Form() {
       <Row>
         <InfoSection className="col-start-2 col-end-3" />
       </Row>
-      {Array.from({ length: addSection }).map((i, index) => (
+      {Array.from({ length: sectionsCount }).map((i, index) => (
         <Row key={index}>
           <GPTGeneratedSection className="col-start-2 col-end-3" />
-          <DeleteButton onClick={handleRemoveSection} />
+          <DeleteButton onClick={onClickDelete} />
         </Row>
       ))}
-      <Row>
-        <AddButton
-          onClick={handleAddSection}
-          className="col-start-3 col-end-4"
-        />
-      </Row>
     </form>
   );
 }
